backups/backup_1751240923/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
black==25.1.0
click==8.2.1
colorama==0.4.6
ecdsa==0.19.1
exceptiongroup==1.3.0
graphql-core==3.2.6
greenlet==3.2.2
h11==0.16.0
idna==3.10
inflect==7.5.0
more-itertools==10.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pyasn1==0.6.1
pydantic-settings==2.9.1
pydantic_core==2.33.2
pyodbc==5.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-jose==3.5.0
python-multipart==0.0.20
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlacodegen==3.0.0
SQLAlchemy==2.0.41
starlette==0.46.2
strawberry-graphql==0.270.4
tomli==2.2.1
typeguard==4.4.2
typing-inspection==0.4.1
typing_extensions==4.13.2
uvicorn==0.34.2
